WIP: Add neo4j migrations dependency and initial migration script
Added a new dependency 'neo4j-migrations-spring-boot-starter' to the project's pom.xml. Created a cypher script for the initial migration to enforce a unique constraint on customer's email. Also, updated application.properties file to include the configuration for migrations, specifying the location to scan for migration scripts and disabling the check for its existence.